+++ CHANGELOG.md
@@ -7,3 +7,10 @@ Changes of v 0.40 to v 0.42 (July 2017)
   development info, and the making of statistical graphics.
 
 We thank to the professor Alex Buerkle his comments to improve ddRADseqTools.
+
+********************************************************************************
+
+Changes of v 0.42 to v 0.43 (October 2017)
+
+* enzyme1 and enzyme parameters can be an identifier or a 
+  restriction site sequence, independently of each other.